fixing resource id and policy endpoints
